pyproject.toml
[tool.poetry]
name = "shaniraq-kz"
version = "0.1.0"
description = ""
authors = ["Dias <saidshabekov@gmail.com>"]
readme = "README.md"
packages = [{include = "shaniraq"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = "^0.23.1"
sqlalchemy = "^2.0.19"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
alembic = "^1.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
